+++ vm/src/vm/test.rs
@@ -52,7 +52,7 @@ pub fn setup() -> (Machine, Block, RcProcess) {
 /// Creates a new instruction.
 pub fn new_instruction(
     ins_type: InstructionType,
-    args: Vec<usize>,
+    args: Vec<u16>,
 ) -> Instruction {
     Instruction::new(ins_type, args, 1)
 }
